// packages/ruig/src/layers/view/application/components/base/model/IStyle.ts
type IStyle = {
  overflow?: string
  background?: string
  color?: string
  border?: string
  display?: string
  position?: string
  margin?: string
  width?: string
  height?: string
  top?: string | number
  right?: string | number
  left?: string | number
  bottom?: string | number
  padding?: string
  transform?: string
  transition?: string
  'border-radius'?: string
  'flex-wrap'?: 'wrap' | 'nowrap' | ''
  'flex-direction'?: 'column' | 'grid' | ''
  'justify-content'?: 'space-around' | 'space-between' | ''
  'z-index'?: string | number
  'max-height'?: string
  'min-height'?: string
  'max-width'?: string
  'min-width'?: string
  'overflow-y'?: string
  'overflow-x'?: string
  'box-shadow'?: string
  'border-color'?: string
  'border-top'?: string
  'border-bottom'?: string
}

export { IStyle }